% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{v_filter}
\alias{v_filter}
\title{Filter a vector}
\usage{
v_filter(.x, .f = v_all_true)
}
\arguments{
\item{.x}{A vector.}

\item{.f}{A single predicate function (or purrr formula) or a logical vector of the same length as .x.}
}
\value{
The vector .x with (possibly) some entries deleted.
}
\description{
Filter a vector
}
